TSに変更
